Reorganization of the library package structure, at petition by @Darklight45.
