+++ src/cu/edu/cujae/graphy/algorithms/CycleDetection.java
@@ -20,7 +20,7 @@ package cu.edu.cujae.graphy.algorithms;
 
 import cu.edu.cujae.graphy.core.Edge;
 import cu.edu.cujae.graphy.core.Graph;
-import cu.edu.cujae.graphy.core.GraphIterator;
+import cu.edu.cujae.graphy.core.iterators.GraphIterator;
 import java.util.Objects;
 import java.util.Set;
 import java.util.TreeSet;
