+++ src/cu/edu/cujae/graphy/core/defaults/DefaultNode.java
@@ -16,8 +16,10 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package cu.edu.cujae.graphy.core;
+package cu.edu.cujae.graphy.core.defaults;
 
+import cu.edu.cujae.graphy.core.Edge;
+import cu.edu.cujae.graphy.core.Node;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
