+++ src/cu/edu/cujae/graphy/core/defaults/DefaultNotDirectedEdgeFactory.java
@@ -1,4 +1,10 @@
-package cu.edu.cujae.graphy.core;
+package cu.edu.cujae.graphy.core.defaults;
+
+import cu.edu.cujae.graphy.core.abstractions.AbstractEdge;
+import cu.edu.cujae.graphy.core.Edge;
+import cu.edu.cujae.graphy.core.EdgeFactory;
+import cu.edu.cujae.graphy.core.Node;
+import cu.edu.cujae.graphy.core.Weight;
 
 public class DefaultNotDirectedEdgeFactory extends DefaultEdgeFactory implements EdgeFactory {
