+++ src/cu/edu/cujae/graphy/core/iterators/GraphIterator.java
@@ -16,8 +16,10 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA 02110-1301  USA
  */
-package cu.edu.cujae.graphy.core;
+package cu.edu.cujae.graphy.core.iterators;
 
+import cu.edu.cujae.graphy.core.Edge;
+import cu.edu.cujae.graphy.core.Node;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.Set;
